+++ app/src/main/java/com/chico/myhomebookkeeping/ui/paymentPackage/moneyMoving/dialogs/SelectMoneyMovingDialog.kt
@@ -64,7 +64,7 @@ class SelectMoneyMovingDialog(
             dateTime.text = fullMoneyMoving.timeStamp.parseTimeFromMillis()
             amount.text = fullMoneyMoving.amount.toString()
             currency.text = fullMoneyMoving.currencyNameValue
-            category.text = fullMoneyMoving.categoryNameValue ?: fullMoneyMoving.paymentTypeName
+            category.text = fullMoneyMoving.categoryDisplayName ?: fullMoneyMoving.paymentTypeName
             cashAccount.text = fullMoneyMoving.cashAccountNameValue
             if (!fullMoneyMoving.description.isNullOrEmpty()) {
                 description.visibility = View.VISIBLE